fix(collision): enable powerup pickup during wave intro
checkCollisions was not called during StateWaveIntro, so powerups
sat on the ground uncollectable until the next wave started. Extract
checkPowerUpCollisions into its own function and call it from
updateWaveIntro. Also migrates game-wide coordinates from float64 to
float32 to match Ebitengine's vector API and adds math32 helpers.

+++ game/event.go
@@ -26,7 +26,7 @@ const (
 // Event is a value type describing something that happened this frame.
 type Event struct {
 	Type   EventType
-	X, Y   float64
-	Value  float64
+	X, Y   float32
+	Value  float32
 	Silent bool // suppress sound (e.g. missile blast covers individual kills)
 }
